Middleware function added
