+++ src/services/session.service.ts
@@ -10,16 +10,16 @@ import * as keys from "../session/keys";
  * @param companyNumber
  */
 const createPromiseToFileSession = async (
-  chSession: Session, companyNumber: string): Promise<IPromiseToFileSession> => {
+  chSession: Session): Promise<IPromiseToFileSession> => {
   const ptfSession: IPromiseToFileSession = {
-    company_number_in_context: companyNumber,
+    company_number: "",
   };
   chSession.appendData(keys.PTF_SESSION, ptfSession);
   await saveSession(chSession);
   return ptfSession;
 };
 
-const updatePTFSessionValue = async (chSession: Session, key: string, value: any): Promise<void> => {
+const updatePromiseToFileSessionValue = async (chSession: Session, key: string, value: any): Promise<void> => {
   const ptfSession = await chSession.data.ptf_session;
   ptfSession[key] = value;
   chSession.appendData(keys.PTF_SESSION, ptfSession);
@@ -35,4 +35,4 @@ const getCompanyNumberInContext = (chSession: Session): string => {
   return chSession.data.ptf_session.company_number_in_context;
 };
 
-export { createPromiseToFileSession, getCompanyNumberInContext, updatePTFSessionValue };
+export { createPromiseToFileSession, getCompanyNumberInContext, updatePromiseToFileSessionValue };
