+++ src/session/middleware/index.ts
@@ -1,27 +1,11 @@
-import Session from "../session";
-import {NextFunction, Request, Response} from "express";
-import {COOKIE_NAME} from "../../properties";
-import * as redisService from "../../services/redis.service";
-import logger from "../../logger";
-
-declare global {
-  namespace Express {
-    interface Request {
-      chSession: Session;
-    }
-  }
-}
+import { NextFunction, Request, Response } from "express";
+import * as keys from "../keys";
+import * as sessionService from "../../services/session.service";
 
 export default async (req: Request, res: Response, next: NextFunction) => {
-  const cookieId = req.cookies[COOKIE_NAME];
-
-  if (cookieId) {
-    logger.info("cookie found, loading session from redis: " + cookieId);
-    req.chSession = await redisService.loadSession(cookieId);
-  } else {
-    logger.info("No cookie found, using blank session");
-    req.chSession = await redisService.loadSession("");
+  const existing = req.chSession.data[keys.PTF_SESSION];
+  if (!existing) {
+    await sessionService.createPromiseToFileSession(req.chSession);
   }
-
   next();
 };
